refactor(swap): extract rune logic to hook
